Move verify child workflow completion recorded to api package
